Deleting Seattle no longer does bad stuff...
No longer tries to access bad memory when printing the cities after Seattle is deleted. Had to comment out the call to rbDeleteFixup in rbDelete however to get this to work. So the bad news is that wie will probably end up switching to a non red-black tree if we can't solve these issure correctly.